kn - formatted column for quarter and added tests
